refactor(playbook): split player chrome and console
