+++ apps/web/src/features/playbook/engine/player/PlaybookPlayer.tsx
@@ -5,8 +5,7 @@ import type { DirectorScript, PlaybookScript } from "../types";
 import { usePlaybookController } from "./usePlaybookController";
 import { PlaybookComposition } from "../composition/PlaybookComposition";
 import { PLAYBOOK_DEFAULTS } from "../../../../shared/config/constants";
-import { useTTS, OPENAI_VOICES, AUTO_VOICE, resolveVoice } from "./useTTS";
-import type { TTSConfig } from "./useTTS";
+import { useTTS, resolveVoice } from "./useTTS";
 import { useKeyboardShortcuts } from "./useKeyboardShortcuts";
 import { resolveNarrationTemplate } from "./resolveNarrationTemplate";
 import { useResolvedScript, type ScriptOverrides } from "./useResolvedScript";
@@ -15,102 +14,17 @@ import { resolveInitialPreviewFrame, resolvePlayerTimelineKey } from "./previewF
 import { CodeHighlightRenderer } from "../renderers/CodeHighlightRenderer";
 import { domainCapability } from "../domainCapabilities";
 import { getParamPanel } from "../param-panels/registry";
-import type { ParamPanelProps } from "../param-panels/types";
 import { hasReplayableAlgorithmParams } from "../param-panels/AlgorithmParamPanel";
 import { resolveDirectorVoiceover } from "../director";
 import { emitNativeEvent } from "../../../../shared/native/emitNativeEvent";
 import { MobileSheet } from "./MobileSheet";
+import { ParamPanelSlot } from "./ParamPanelSlot";
+import { PlaybookLearningConsole } from "./PlaybookLearningConsole";
+import { PlayerSettingsPopover } from "./PlayerSettingsPopover";
+import { ExportSVG, MoreSVG, SettingsSVG, TopbarFoldIcon } from "./PlaybookPlayerIcons";
 import { PlaybookPortraitShell, type MobileTabKey } from "./PlaybookPortraitShell";
 import { clipCodeOverlay } from "./mobileCodeOverlay";
-
-// ── ParamPanelSlot (static component — resolves domain panel from registry) ──
-
-function ParamPanelSlot({ domain, ...props }: ParamPanelProps & { domain: string }) {
-  const Panel = getParamPanel(domain);
-  if (!Panel) return null;
-  return React.createElement(Panel, props);
-}
-
-// ── SVG icons ──────────────────────────────────────────────────────────────
-
-const SpeakerOnSVG = () => (
-  <svg width="14" height="14" viewBox="0 0 24 24" fill="none"
-       stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-    <polygon points="11 5 6 9 2 9 2 15 6 15 11 19 11 5"/>
-    <path d="M15.54 8.46a5 5 0 0 1 0 7.07"/>
-    <path d="M19.07 4.93a10 10 0 0 1 0 14.14"/>
-  </svg>
-);
-
-const SpeakerOffSVG = () => (
-  <svg width="14" height="14" viewBox="0 0 24 24" fill="none"
-       stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-    <polygon points="11 5 6 9 2 9 2 15 6 15 11 19 11 5"/>
-    <line x1="23" y1="9" x2="17" y2="15"/>
-    <line x1="17" y1="9" x2="23" y2="15"/>
-  </svg>
-);
-
-const SettingsSVG = () => (
-  <svg width="13" height="13" viewBox="0 0 24 24" fill="none"
-       stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-    <circle cx="12" cy="12" r="3"/>
-    <path d="M19.4 15a1.65 1.65 0 0 0 .33 1.82l.06.06a2 2 0 0 1-2.83 2.83l-.06-.06
-             a1.65 1.65 0 0 0-1.82-.33 1.65 1.65 0 0 0-1 1.51V21a2 2 0 0 1-4 0v-.09
-             A1.65 1.65 0 0 0 9 19.4a1.65 1.65 0 0 0-1.82.33l-.06.06a2 2 0 0 1-2.83-2.83
-             l.06-.06A1.65 1.65 0 0 0 4.68 15a1.65 1.65 0 0 0-1.51-1H3a2 2 0 0 1 0-4h.09
-             A1.65 1.65 0 0 0 4.6 9a1.65 1.65 0 0 0-.33-1.82l-.06-.06a2 2 0 0 1 2.83-2.83
-             l.06.06A1.65 1.65 0 0 0 9 4.68a1.65 1.65 0 0 0 1-1.51V3a2 2 0 0 1 4 0v.09
-             a1.65 1.65 0 0 0 1 1.51 1.65 1.65 0 0 0 1.82-.33l.06-.06a2 2 0 0 1 2.83 2.83
-             l-.06.06A1.65 1.65 0 0 0 19.4 9a1.65 1.65 0 0 0 1.51 1H21a2 2 0 0 1 0 4h-.09
-             a1.65 1.65 0 0 0-1.51 1z"/>
-  </svg>
-);
-
-const ExportSVG = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none"
-       stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"
-       aria-hidden="true">
-    <path d="M12 3v12" />
-    <path d="m7 10 5 5 5-5" />
-    <path d="M5 21h14" />
-  </svg>
-);
-
-const MoreSVG = () => (
-  <svg width="15" height="15" viewBox="0 0 24 24" fill="none"
-       stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"
-       aria-hidden="true">
-    <circle cx="5" cy="12" r="1" />
-    <circle cx="12" cy="12" r="1" />
-    <circle cx="19" cy="12" r="1" />
-  </svg>
-);
-
-function TopbarFoldIcon({ collapsed }: { collapsed: boolean }) {
-  return (
-    <svg
-      className="playbook-player__chrome-toggle-icon"
-      data-testid={
-        collapsed ? "topbar-toggle-icon-expand" : "topbar-toggle-icon-collapse"
-      }
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="1.55"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      aria-hidden="true"
-    >
-      <path className="playbook-player__chrome-toggle-bar" d="M7 7.5h10" />
-      {collapsed ? (
-        <path className="playbook-player__chrome-toggle-chevron" d="m7.5 11 4.5 4.5 4.5-4.5" />
-      ) : (
-        <path className="playbook-player__chrome-toggle-chevron" d="m7.5 14 4.5-4.5 4.5 4.5" />
-      )}
-    </svg>
-  );
-}
+import { SPEED_STEPS } from "./playbackRates";
 
 export type PlaybookLayoutMode = "desktop" | "portrait";
 
@@ -138,268 +52,6 @@ function useAutoLayoutMode(layoutMode?: PlaybookLayoutMode): PlaybookLayoutMode 
 
   return layoutMode ?? autoLayoutMode;
 }
-
-// ── Player Settings Popover ────────────────────────────────────────────────
-
-interface PlayerSettingsPopoverProps {
-  playbackRate: number;
-  onPlaybackRateChange: (rate: number) => void;
-  stepThrough: boolean;
-  onStepThroughChange: (next: boolean) => void;
-  ttsEnabled: boolean;
-  ttsSupported: boolean;
-  onToggleTTS: () => void;
-  config: TTSConfig;
-  onUpdate: (patch: Partial<TTSConfig>) => void;
-  onClose: () => void;
-  isDark: boolean;
-  onPreview: (voice: string, sampleText: string) => void;
-}
-
-const SAMPLE_TEXT_DEFAULT = "你好，这是一段试听文字。Hello, this is a preview.";
-const SPEED_STEPS = [0.5, 0.75, 1, 1.25, 1.5, 2];
-
-const PlayerSettingsPopover: React.FC<PlayerSettingsPopoverProps> = ({
-  playbackRate,
-  onPlaybackRateChange,
-  stepThrough,
-  onStepThroughChange,
-  ttsEnabled,
-  ttsSupported,
-  onToggleTTS,
-  config,
-  onUpdate,
-  onClose,
-  isDark,
-  onPreview,
-}) => {
-  const popoverRef = useRef<HTMLDivElement>(null);
-  const [sampleText, setSampleText] = useState(SAMPLE_TEXT_DEFAULT);
-
-  useEffect(() => {
-    const handler = (e: MouseEvent) => {
-      const target = e.target as Element | null;
-      if (target?.closest(".playbook-player__settings-anchor")) return;
-      if (popoverRef.current && !popoverRef.current.contains(e.target as Node)) {
-        onClose();
-      }
-    };
-    document.addEventListener("mousedown", handler);
-    return () => document.removeEventListener("mousedown", handler);
-  }, [onClose]);
-
-  const bg = isDark ? "#161b22" : "#ffffff";
-  const border = isDark ? "#30363d" : "#d0d7de";
-  const text = isDark ? "#c9d1d9" : "#24292f";
-  const muted = isDark ? "#8b949e" : "#6e7781";
-  const inputBg = isDark ? "#0d1117" : "#f6f8fa";
-  const accent = isDark ? "#4de8b0" : "#00896e";
-
-  return (
-    <div
-      ref={popoverRef}
-      className="playbook-player__settings-popover"
-      style={{
-        "--player-settings-bg": bg,
-        "--player-settings-border": border,
-        "--player-settings-text": text,
-        "--player-settings-muted": muted,
-        "--player-settings-input": inputBg,
-        "--player-settings-accent": accent,
-      } as React.CSSProperties}
-    >
-      <div className="playbook-player__settings-section">
-        <div className="playbook-player__settings-label">播放速度</div>
-        <div className="playbook-player__settings-speed-row">
-          {SPEED_STEPS.map((s) => (
-            <button
-              key={s}
-              type="button"
-              onClick={() => onPlaybackRateChange(s)}
-              className={playbackRate === s ? "is-active" : ""}
-            >
-              {s}×
-            </button>
-          ))}
-        </div>
-      </div>
-
-      <div className="playbook-player__settings-section">
-        <div className="playbook-player__settings-row">
-          <span>播放模式</span>
-          <button
-            type="button"
-            className={`playbook-player__settings-toggle${stepThrough ? " is-active" : ""}`}
-            onClick={() => onStepThroughChange(!stepThrough)}
-          >
-            {stepThrough ? "步进" : "连播"}
-          </button>
-        </div>
-        <div className="playbook-player__settings-row">
-          <span>语音朗读</span>
-          <button
-            type="button"
-            className={`playbook-player__settings-toggle${ttsEnabled ? " is-active" : ""}`}
-            onClick={onToggleTTS}
-            disabled={!ttsSupported}
-          >
-            {ttsEnabled ? <SpeakerOnSVG /> : <SpeakerOffSVG />}
-            {ttsEnabled ? "开启" : "关闭"}
-          </button>
-        </div>
-      </div>
-
-      <div className="playbook-player__settings-section">
-        <div className="playbook-player__settings-label">语音后端</div>
-        <div className="playbook-player__settings-choice-row">
-          {(["system", "openai"] as const).map((b) => (
-            <button
-              key={b}
-              type="button"
-              onClick={() => onUpdate({ backend: b })}
-              className={config.backend === b ? "is-active" : ""}
-            >
-              {b === "system" ? "系统语音" : "OpenAI API"}
-            </button>
-          ))}
-        </div>
-      </div>
-
-      {/* OpenAI backend now routes through the server-side proxy
-         (POST /api/v1/tts/speech) so the player never stores an API key.
-         The Base URL / Model / Key inputs that used to live here are gone
-         on purpose — issue #40. */}
-      {config.backend === "openai" && (
-        <div
-          className="playbook-player__settings-note"
-          style={{
-            border: `1px dashed ${border}`,
-          }}
-        >
-          通过服务端代理调用，API Key 由管理员在后端配置（METAVIEW_TTS_API_KEY）。
-        </div>
-      )}
-
-      {/* Voice picker (OpenAI only) */}
-      {config.backend === "openai" && (
-        <div>
-          <div className="playbook-player__settings-label">
-            音色 <span style={{ color: text }}>{config.voice === AUTO_VOICE ? "跟随学科" : config.voice}</span>
-          </div>
-          <div style={{ display: "flex", flexDirection: "column", gap: 4, maxHeight: 180, overflowY: "auto" }}>
-            <button
-              type="button"
-              onClick={() => onUpdate({ voice: AUTO_VOICE })}
-              style={{
-                display: "flex",
-                alignItems: "center",
-                justifyContent: "space-between",
-                padding: "5px 8px",
-                borderRadius: 5,
-                border: `1px solid ${config.voice === AUTO_VOICE ? accent : border}`,
-                background: config.voice === AUTO_VOICE ? `${accent}18` : "transparent",
-                color: config.voice === AUTO_VOICE ? accent : text,
-                fontSize: 12,
-                cursor: "pointer",
-                textAlign: "left",
-              }}
-            >
-              <span>跟随学科自动推荐</span>
-            </button>
-            {OPENAI_VOICES.map((v) => (
-              <div
-                key={v.id}
-                style={{
-                  display: "flex",
-                  alignItems: "center",
-                  gap: 6,
-                  padding: "4px 8px",
-                  borderRadius: 5,
-                  border: `1px solid ${config.voice === v.id ? accent : border}`,
-                  background: config.voice === v.id ? `${accent}18` : "transparent",
-                }}
-              >
-                <button
-                  type="button"
-                  onClick={() => onUpdate({ voice: v.id })}
-                  style={{
-                    flex: 1,
-                    background: "transparent",
-                    border: "none",
-                    color: config.voice === v.id ? accent : text,
-                    fontSize: 12,
-                    fontWeight: config.voice === v.id ? 600 : 400,
-                    textAlign: "left",
-                    cursor: "pointer",
-                    padding: 0,
-                  }}
-                  title={v.description}
-                >
-                  {v.label}
-                  <span style={{ color: muted, fontWeight: 400, marginLeft: 6 }}>{v.description}</span>
-                </button>
-                <button
-                  type="button"
-                  onClick={() => onPreview(v.id, sampleText)}
-                  style={{
-                    background: "transparent",
-                    border: `1px solid ${border}`,
-                    borderRadius: 4,
-                    color: muted,
-                    fontSize: 11,
-                    padding: "2px 6px",
-                    cursor: "pointer",
-                  }}
-                >
-                  ⏵ 试听
-                </button>
-              </div>
-            ))}
-          </div>
-          <div style={{ marginTop: 6 }}>
-            <div style={{ fontSize: 11, color: muted, marginBottom: 3 }}>试听文字</div>
-            <input
-              type="text"
-              value={sampleText}
-              onChange={(e) => setSampleText(e.target.value)}
-              placeholder={SAMPLE_TEXT_DEFAULT}
-              style={{
-                width: "100%",
-                padding: "5px 8px",
-                background: inputBg,
-                border: `1px solid ${border}`,
-                borderRadius: 5,
-                color: text,
-                fontSize: 12,
-                outline: "none",
-                boxSizing: "border-box",
-              }}
-            />
-          </div>
-        </div>
-      )}
-
-      <div className="playbook-player__settings-section">
-        <div className="playbook-player__settings-label">
-          语速 <span style={{ color: text }}>{config.rate.toFixed(1)}×</span>
-        </div>
-        <input
-          type="range"
-          min="0.5"
-          max="2.0"
-          step="0.1"
-          value={config.rate}
-          onChange={(e) => onUpdate({ rate: parseFloat(e.target.value) })}
-          className="playbook-player__settings-range"
-        />
-      </div>
-
-      <div className="playbook-player__settings-footnote">
-        设置存储在本地浏览器中
-      </div>
-    </div>
-  );
-};
 
 // ── Main component ─────────────────────────────────────────────────────────
 
@@ -889,66 +541,18 @@ export const PlaybookPlayer: React.FC<PlaybookPlayerProps> = ({
       )}
 
       {showLearningConsole && !isPortraitLayout && (
-        <aside className="playbook-player__console" aria-label="Learning console">
-          {showCodePanelSlot && (
-            <section className="playbook-player__console-card playbook-player__code-card">
-              <div className="playbook-player__console-head">
-                <span>Code Sync</span>
-                <small>{codeOverlay?.language ?? "source"}</small>
-              </div>
-              <div className="playbook-player__code-body">
-                {codeOverlay ? (
-                  <CodeHighlightRenderer overlay={codeOverlay} theme={theme} />
-                ) : (
-                  <div className="playbook-player__code-empty">
-                    <span>{"</>"}</span>
-                    <p>Code highlights will sync here.</p>
-                  </div>
-                )}
-              </div>
-            </section>
-          )}
-
-          {hasDomainPanel && (
-            <section className="playbook-player__console-card playbook-player__params-card">
-              <div className="playbook-player__console-head">
-                <span>Params</span>
-                <small>{baseScript.domain}</small>
-              </div>
-              <div className="playbook-player__param-body">
-                <ParamPanelSlot
-                  domain={baseScript.domain}
-                  script={baseScript}
-                  overrides={overrides}
-                  onOverridesChange={setOverrides}
-                  isDark={theme === "dark"}
-                />
-              </div>
-            </section>
-          )}
-
-          {followupSlot && (
-            <section className="playbook-player__console-card playbook-player__follow-card">
-              <div className="playbook-player__console-head">
-                <span>Follow-up</span>
-                <small>current step</small>
-              </div>
-              <div className="playbook-player__follow-body">{followupSlot}</div>
-            </section>
-          )}
-
-          {relatedSlot ? (
-            <section className="playbook-player__related-card" aria-label="Related study context">
-              {relatedSlot}
-            </section>
-          ) : (
-            <section className="playbook-player__related-row" aria-label="Related study context">
-              <span>Related</span>
-              <strong>{script.algorithm_id ?? "Study variants"}</strong>
-              <small>›</small>
-            </section>
-          )}
-        </aside>
+        <PlaybookLearningConsole
+          showCodePanelSlot={showCodePanelSlot}
+          codeOverlay={codeOverlay}
+          theme={theme}
+          hasDomainPanel={hasDomainPanel}
+          baseScript={baseScript}
+          overrides={overrides}
+          onOverridesChange={setOverrides}
+          followupSlot={followupSlot}
+          relatedSlot={relatedSlot}
+          relatedAlgorithmId={script.algorithm_id}
+        />
       )}
 
       {showMobileConsole && mobileSheet && (
