+++ apps/web/src/features/playbook/engine/player/playbackRates.ts
@@ -1,0 +1,1 @@
+export const SPEED_STEPS = [0.5, 0.75, 1, 1.25, 1.5, 2];
